Add button for importing a JSON timeline
This uses a standard HTML file input, but intercepts the event instead
of providing the normal POST action. We can use the javascript File
object to read the contents without it having to leave the user's
browser.

We've wrapped the parsing and import in a try/catch in case the user
uploads invalid JSON.