Migrate storage.sync to storage.local for timeline
Migrate from storage.sync to storage.local for the timeline because of
the small storage limit in storage.sync. The migration will occur when
the extension loads. The migration will attempt to move any sync
timeline to local storage. If there's already a local timeline, it will
merge and deduplicate the two timelines. The sync timeline will only be
removed if the local timeline is the expected timeline.

In addition, a notification will be fired whenever the storage limit
error is thrown suggesting that users export and reset their stats.

Related issue #82